+++ src/test/java/com/ericsson/bss/cassandra/ecaudit/integration/ITFunctionsAudit.java
@@ -70,11 +70,27 @@ public class ITFunctionsAudit
             cassandraSession.execute(new SimpleStatement(
             "CREATE ROLE superfunc WITH PASSWORD = 'secret' AND LOGIN = true AND SUPERUSER = true"));
             cassandraSession.execute(new SimpleStatement(
-            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_all' : 'roles'}"));
+            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_execute' : 'connections'}"));
             cassandraSession.execute(new SimpleStatement(
-            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_all' : 'data'}"));
+            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_create' : 'roles'}"));
             cassandraSession.execute(new SimpleStatement(
-            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_all' : 'functions'}"));
+            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_alter' : 'roles'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_drop' : 'roles'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_authorize' : 'roles'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_create' : 'data'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_alter' : 'data'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_drop' : 'data'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_select' : 'data'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_modify' : 'data'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superfunc WITH OPTIONS = { 'grant_audit_whitelist_for_create' : 'functions'}"));
         }
 
         superCluster = cdt.createCluster("superfunc", "secret");
@@ -315,7 +331,7 @@ public class ITFunctionsAudit
         givenFinalStateFunction(keyspace, aggregate + "FSF");
         superSession.execute(new SimpleStatement(
         "CREATE AGGREGATE IF NOT EXISTS " + keyspace + "." + aggregate + " (int) " +
-        "SFUNC " + aggregate + "SF "+
+        "SFUNC " + aggregate + "SF " +
         "STYPE tuple<int,bigint> " +
         "FINALFUNC " + aggregate + "FSF " +
         "INITCOND (0,0)"));
@@ -327,11 +343,11 @@ public class ITFunctionsAudit
         superSession.execute(new SimpleStatement(
         "CREATE ROLE " + username + " WITH PASSWORD = 'secret' AND LOGIN = true AND SUPERUSER = true"));
         superSession.execute(new SimpleStatement(
-        "ALTER ROLE " + username + " WITH OPTIONS = { 'grant_audit_whitelist_for_all'  : 'data/system' }"));
+        "ALTER ROLE " + username + " WITH OPTIONS = { 'grant_audit_whitelist_for_execute'  : 'connections' }"));
         superSession.execute(new SimpleStatement(
-        "ALTER ROLE " + username + " WITH OPTIONS = { 'grant_audit_whitelist_for_all'  : 'data/system_schema' }"));
+        "ALTER ROLE " + username + " WITH OPTIONS = { 'grant_audit_whitelist_for_select'  : 'data/system' }"));
         superSession.execute(new SimpleStatement(
-        "ALTER ROLE " + username + " WITH OPTIONS = { 'grant_audit_whitelist_for_all'  : 'connections' }"));
+        "ALTER ROLE " + username + " WITH OPTIONS = { 'grant_audit_whitelist_for_select'  : 'data/system_schema' }"));
         return username;
     }
 
@@ -342,7 +358,7 @@ public class ITFunctionsAudit
 
     private void whenRoleIsWhitelistedForOperationOnResource(String username, String operation, String resource)
     {
-        superSession.execute("ALTER ROLE " + username + " WITH OPTIONS = {'grant_audit_whitelist_for_all' : '" + resource + "'}");
+        superSession.execute("ALTER ROLE " + username + " WITH OPTIONS = {'grant_audit_whitelist_for_" + operation + "' : '" + resource + "'}");
     }
 
     private void thenAuditLogContainNothingForUser()