+++ src/test/java/com/ericsson/bss/cassandra/ecaudit/integration/ITRolesAudit.java
@@ -70,9 +70,23 @@ public class ITRolesAudit
             cassandraSession.execute(new SimpleStatement(
             "CREATE ROLE superrole WITH PASSWORD = 'secret' AND LOGIN = true AND SUPERUSER = true"));
             cassandraSession.execute(new SimpleStatement(
-            "ALTER ROLE superrole WITH OPTIONS = { 'grant_audit_whitelist_for_all' : 'roles'}"));
+            "ALTER ROLE superrole WITH OPTIONS = { 'grant_audit_whitelist_for_create' : 'roles'}"));
             cassandraSession.execute(new SimpleStatement(
-            "ALTER ROLE superrole WITH OPTIONS = { 'grant_audit_whitelist_for_all' : 'data'}"));
+            "ALTER ROLE superrole WITH OPTIONS = { 'grant_audit_whitelist_for_alter' : 'roles'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superrole WITH OPTIONS = { 'grant_audit_whitelist_for_drop' : 'roles'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superrole WITH OPTIONS = { 'grant_audit_whitelist_for_authorize' : 'roles'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superrole WITH OPTIONS = { 'grant_audit_whitelist_for_create' : 'data'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superrole WITH OPTIONS = { 'grant_audit_whitelist_for_alter' : 'data'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superrole WITH OPTIONS = { 'grant_audit_whitelist_for_drop' : 'data'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superrole WITH OPTIONS = { 'grant_audit_whitelist_for_select' : 'data'}"));
+            cassandraSession.execute(new SimpleStatement(
+            "ALTER ROLE superrole WITH OPTIONS = { 'grant_audit_whitelist_for_modify' : 'data'}"));
         }
 
         superCluster = cdt.createCluster("superrole", "secret");
@@ -304,11 +318,11 @@ public class ITRolesAudit
         superSession.execute(new SimpleStatement(
         "CREATE ROLE " + username + " WITH PASSWORD = 'secret' AND LOGIN = true AND SUPERUSER = true"));
         superSession.execute(new SimpleStatement(
-        "ALTER ROLE " + username + " WITH OPTIONS = { 'grant_audit_whitelist_for_all'  : 'connections' }"));
+        "ALTER ROLE " + username + " WITH OPTIONS = { 'grant_audit_whitelist_for_execute'  : 'connections' }"));
         superSession.execute(new SimpleStatement(
-        "ALTER ROLE " + username + " WITH OPTIONS = { 'grant_audit_whitelist_for_all'  : 'data/system' }"));
+        "ALTER ROLE " + username + " WITH OPTIONS = { 'grant_audit_whitelist_for_select'  : 'data/system' }"));
         superSession.execute(new SimpleStatement(
-        "ALTER ROLE " + username + " WITH OPTIONS = { 'grant_audit_whitelist_for_all'  : 'data/system_schema' }"));
+        "ALTER ROLE " + username + " WITH OPTIONS = { 'grant_audit_whitelist_for_select'  : 'data/system_schema' }"));
         return username;
     }
 
@@ -325,7 +339,7 @@ public class ITRolesAudit
 
     private void whenRoleIsWhitelistedForOperationOnResource(String username, String operation, String resource)
     {
-        superSession.execute("ALTER ROLE " + username + " WITH OPTIONS = {'grant_audit_whitelist_for_all' : '" + resource + "'}");
+        superSession.execute("ALTER ROLE " + username + " WITH OPTIONS = {'grant_audit_whitelist_for_" + operation + "' : '" + resource + "'}");
     }
 
     private void thenAuditLogContainNothingForUser()